+++ Utilities/DebugConsole/changelog.md
@@ -1,5 +1,9 @@
 Changelog of Debug Console, a Phoenix Point Mod by Sheepy
 
+# Version 7, in development
+
+* New api "console.shell start" and "console.shell stop" to create a console shell with caller function.
+
 # Version 6.1, 2020-07-10
 
 * Fix: Handle TypeLoadException when scanning a class for command methods.